+++ packages/host-api/src/hostApi.ts
@@ -5,7 +5,7 @@ import type { Codec, CodecType } from 'scale-ts';
 import { extractErrorMessage } from './helpers.js';
 import { GenericError } from './protocol/commonCodecs.js';
 import type { HostApiProtocol, VersionedProtocolRequest, VersionedProtocolSubscription } from './protocol/impl.js';
-import { CreateProofErr, GetUserIdErr, LoginErr, RequestCredentialsErr } from './protocol/v1/accounts.js';
+import { CreateProofErr, GetUserIdErr, LoginErr, RequestCredentialsErr, SignVrfErr } from './protocol/v1/accounts.js';
 import { ChatBotRegistrationErr, ChatMessagePostingErr, ChatRoomRegistrationErr } from './protocol/v1/chat.js';
 import { CreateTransactionErr } from './protocol/v1/createTransaction.js';
 import { DeriveEntropyErr } from './protocol/v1/deriveEntropy.js';
@@ -187,6 +187,13 @@ export function createHostApi(transport: Transport): HostApi {
       return makeRequest(transport.request('host_account_create_proof', payload), reason => ({
         tag: payload.tag,
         value: new CreateProofErr.Unknown({ reason }),
+      }));
+    },
+
+    accountSignVrf(payload) {
+      return makeRequest(transport.request('host_account_sign_vrf', payload), reason => ({
+        tag: payload.tag,
+        value: new SignVrfErr.Unknown({ reason }),
       }));
     },