+++ packages/host-api/src/index.ts
@@ -54,7 +54,10 @@ export {
   ProductAccountId,
   RequestCredentialsErr,
   RingLocation,
+  SignVrfErr,
   UserIdentity,
+  VrfSignature,
+  VrfTranscriptItem,
 } from './protocol/v1/accounts.js';
 export {
   ChatActionPayload,